requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
supabase==2.3.4
python-dotenv==1.0.0
lxml==4.9.3
fake-useragent==1.4.0
openai==1.3.7
